submit std logs to minio on finish and error

+++ dataquality/core/_config.py
@@ -14,7 +14,6 @@ from requests.exceptions import ConnectionError as ReqConnectionError
 from dataquality import __version__ as dq_version
 from dataquality.exceptions import GalileoException
 from dataquality.schemas.task_type import TaskType
-from dataquality.utils.helpers import GALILEO_DISABLED
 
 
 class GalileoConfigVars(str, Enum):
@@ -143,7 +142,7 @@ def _check_console_url() -> None:
 
 
 def set_config() -> Config:
-    if os.getenv(GALILEO_DISABLED):
+    if os.getenv("GALILEO_DISABLED"):
         return Config(api_url="", minio_url="")
     _check_console_url()
     if not os.path.isdir(ConfigData.DEFAULT_GALILEO_CONFIG_DIR.value):
@@ -188,7 +187,8 @@ def set_config() -> Config:
 
 def reset_config() -> Config:
     """Wipe the config file and reconfigure"""
-    os.remove(ConfigData.DEFAULT_GALILEO_CONFIG_FILE.value)
+    if os.path.isfile(ConfigData.DEFAULT_GALILEO_CONFIG_FILE.value):
+        os.remove(ConfigData.DEFAULT_GALILEO_CONFIG_FILE.value)
     return set_config()
